+++ src/control_plane/shared_infrastructure/rabbitmq/mod.rs
@@ -6,6 +6,7 @@ pub(crate) use rabbitmq_project_definition::RabbitMqProjectDefinition;
 pub(crate) use rabbitmq_project_resources::RabbitMqProjectResources;
 pub(crate) use rabbitmq_shared_instance_plan::RabbitMqSharedInstancePlan;
 pub(crate) use rabbitmq_shared_instance_plan_options::RabbitMqSharedInstancePlanOptions;
+pub(crate) use reload_rabbitmq_definitions::reload_rabbitmq_definitions;
 pub(crate) use store_rabbitmq_definitions::store_rabbitmq_definitions;
 pub(crate) use stored_rabbitmq_paths::StoredRabbitMqPaths;
 
@@ -17,5 +18,6 @@ mod rabbitmq_project_definition;
 mod rabbitmq_project_resources;
 mod rabbitmq_shared_instance_plan;
 mod rabbitmq_shared_instance_plan_options;
+mod reload_rabbitmq_definitions;
 mod store_rabbitmq_definitions;
 mod stored_rabbitmq_paths;